@@ -2,7 +2,7 @@ package org.happymtb.unofficial;
 
 import org.happymtb.unofficial.fragment.PostsListFragment;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.MenuItem;
 
 public class PostsActivity extends AppCompatActivity {
